@@ -23,6 +23,7 @@ use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseDeclaredClass
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseSignatureHelper;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\DocblockCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseSubscriptCompletor;
+use Phpactor\Completion\Bridge\WorseReflection\Completor\ContextSensitiveCompletor;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\ClassFormatter;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\ConstantFormatter;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\EnumCaseFormatter;
@@ -61,6 +62,7 @@ use Phpactor\MapResolver\Resolver;
 use Phpactor\Container\Container;
 use Phpactor\ReferenceFinder\NameSearcher;
 use Phpactor\WorseReflection\Core\Reflector\SourceCodeReflector;
+use Phpactor\WorseReflection\Reflector;
 use RuntimeException;
 
 class CompletionWorseExtension implements Extension
@@ -285,8 +287,8 @@ class CompletionWorseExtension implements Extension
             'imported_names' => [
                 'Completion for names imported into the current namespace',
                 function (Container $container) {
-                    return new ImportedNameCompletor(
-                    );
+                    return $this->contextCompletor($container, new ImportedNameCompletor(
+                    ));
                 },
             ],
             'worse_parameter' => [
@@ -383,13 +385,13 @@ class CompletionWorseExtension implements Extension
             'expression_name_search' => [
                 'Completion for class names, constants and functions at expression positions that are located in the index',
                 function (Container $container) {
-                    return $this->limitCompletor($container, new ExpressionNameCompletor(
+                    return $this->contextCompletor($container, $this->limitCompletor($container, new ExpressionNameCompletor(
                         $container->get(NameSearcher::class),
                         new ObjectFormatter(
                             $container->get(self::SERVICE_COMPLETION_WORSE_SNIPPET_FORMATTERS)
                         ),
                         $container->get(DocumentPrioritizer::class)
-                    ));
+                    )));
                 },
             ],
             'use' => [
@@ -420,7 +422,7 @@ class CompletionWorseExtension implements Extension
                 },
             ],
             'type' => [
-                'Completion for types',
+                'Completion for scalar types',
                 function (Container $container) {
                     return $this->limitCompletor($container, new TypeCompletor(
                         $container->get(TypeSuggestionProvider::class)
@@ -466,5 +468,13 @@ class CompletionWorseExtension implements Extension
         $limit = $container->parameter(self::PARAM_CLASS_COMPLETOR_LIMIT)->int();
 
         return new LimitingCompletor($completor, $limit);
+    }
+
+    private function contextCompletor(Container $container, TolerantCompletor $tolerantCompletor): TolerantCompletor
+    {
+        return new ContextSensitiveCompletor(
+            $tolerantCompletor,
+            $container->expect(WorseReflectionExtension::SERVICE_REFLECTOR, Reflector::class)
+        );
     }
 }
