@@ -134,10 +134,19 @@ class CompletionHandler implements Handler, CanRegisterCapabilities
      */
     public function resolveItem(RequestMessage $request): Promise
     {
+        /** @phpstan-ignore-next-line */
         return call(function () use ($request) {
+            /** @phpstan-ignore-next-line */
             $item = CompletionItem::fromArray($request->params);
-            $item = $this->resolve[$item->data]($item);
-            return $item;
+
+            if (!(is_string($item->data) || is_int($item->data)) || !array_key_exists($item->data, $this->resolve)) {
+                return $item;
+            }
+            /** @phpstan-ignore-next-line - shouldn't happen but playing safe */
+            if (null === $this->resolve[$item->data]) {
+                return $item;
+            }
+            return $this->resolve[$item->data]($item);
         });
     }
 
@@ -189,7 +198,7 @@ class CompletionHandler implements Handler, CanRegisterCapabilities
         Suggestion $suggestion,
         CompletionParams $params
     ): NameImporterResult {
-        $suggestionNameImport = $suggestion->fqn();
+        $suggestionNameImport = $suggestion->nameImport();
 
         if (!$suggestionNameImport) {
             return NameImporterResult::createEmptyResult();
@@ -239,7 +248,7 @@ class CompletionHandler implements Handler, CanRegisterCapabilities
     private function formatShortDescription(Suggestion $suggestion): string
     {
         $prefix = '';
-        if ($suggestion->classImport()) {
+        if ($suggestion->nameImport()) {
             $prefix = '↓ ';
         }
 
